backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
python-dotenv==1.0.0
python-multipart==0.0.20
stripe==12.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
bleach==6.2.0
pytest==8.3.4
pytest-asyncio==1.0.0
httpx==0.28.1
pytest-cov==6.0.0
black==25.1.0
isort==6.0.1 